% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotop.R
\name{html_dependencies_gotop}
\alias{html_dependencies_gotop}
\title{gotop dependencies}
\usage{
html_dependencies_gotop(use_cdn = TRUE)
}
\arguments{
\item{use_cdn}{boolean. TRUE to use a Content Delivery Network (CDN), true by default, or false to use locally stored files.}
}
\description{
Get gotop html dependencies.
}
